refactor + ApiExceptionHandler

+++ src/main/java/com/library/webapp/person/PersonController.java
@@ -1,8 +1,8 @@
-package com.library.webapp.person.controller;
+package com.library.webapp.person;
 
+import com.library.webapp.rental.RentalAlreadyExistsException;
 import com.library.webapp.exception.ResourceNotFoundException;
-import com.library.webapp.person.model.Person;
-import com.library.webapp.person.service.PersonService;
+import com.library.webapp.rental.RentalService;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
@@ -15,9 +15,11 @@ import java.util.Optional;
 @RequestMapping("/api/person")
 public class PersonController {
     private PersonService personService;
+    private RentalService rentalService;
 
-    public PersonController(PersonService personService) {
+    public PersonController(PersonService personService, RentalService rentalService) {
         this.personService = personService;
+        this.rentalService = rentalService;
     }
 
     @GetMapping
@@ -50,14 +52,15 @@ public class PersonController {
             return personService.save(p);
         }).orElseThrow(() -> new ResourceNotFoundException("Person not found"));
     }
+
     @DeleteMapping(value = "/{id}")
     public ResponseEntity<?> deletePerson(@PathVariable Long id) {
+        if (rentalService.countRentalsOfPerson(id) > 0) {
+            throw new RentalAlreadyExistsException("Cannot be deleted because reader has unreturned books!", HttpStatus.INTERNAL_SERVER_ERROR);
+        }
         return personService.findById(id).map(p -> {
             personService.delete(p);
             return ResponseEntity.ok().build();
         }).orElseThrow(() -> new ResourceNotFoundException("Person not found"));
     }
-
-
-
 }
